src/app/tasks/celery_app.py
from celery import Celery
from app.core.config import settings

celery_app = Celery(
    "ocr_worker",
    broker=settings.REDIS_URL,
    backend=settings.REDIS_URL,
    include=["app.tasks.ocr_tasks"],
)

celery_app.conf.update(
    task_serializer="json",
    accept_content=["json"],
    result_serializer="json",
    timezone="UTC",
    enable_utc=True,
    task_track_started=True,
    task_acks_late=True,
    worker_prefetch_multiplier=1,  # One task at a time per worker (OCR is heavy)
    task_soft_time_limit=600,      # 10 minutes soft limit
    task_time_limit=660,           # 11 minutes hard limit
)

