src/requirements.txt
# ── FastAPI core ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Pydantic / Settings ───────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── AI / OCR ─────────────────────────────────────────────────────────────────
torch>=2.3.0
transformers>=5.0.0
Pillow>=10.3.0
pypdfium2>=4.30.0

# ── PDF (digital) to Markdown ─────────────────────────────────────────────────
pymupdf4llm>=0.0.17
pymupdf>=1.24.0

# ── YouTube transcript ────────────────────────────────────────────────────────
youtube-transcript-api>=0.6.2

# ── File validation ───────────────────────────────────────────────────────────
python-magic>=0.4.27

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# ── Database (PostgreSQL) ─────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0                # PostgreSQL async driver
aiosqlite>=0.20.0              # SQLite async driver (used in tests only)
alembic>=1.13.0

# ── Not used in MVP (kept for future reference) ───────────────────────────────
# aiosqlite>=0.20.0            # SQLite async driver (dev only)
# celery[redis]>=5.4.0
# redis>=5.0.0
